requirements.txt
discord.py
python-dotenv
pillow
pytesseract
opencv-python
imagehash
asyncpg
opencv-python